Used matrix multiplication as the dummy simulation
The nx option is used as the matrix size and the KERNELDIR
environment variable is used for locating kernel files.

For example, the following command

  $ KERNELDIR=/disk1/gpgpu/gpu-eps-hav/sim/src build/bin/main --nx 3

computes the product of two random 3x3 matrices.

NOTE: The computation is currently hardcoded to execute on the CPU device,
but this can easily be changed so that it runs on the GPU device instead.
See DummySim::execNextStep() in dummysim.cpp.