Simplified code in attempt to debug kernel compilation problem
Currently the dummy sim program fails in an unstable way when compiling the kernel.
When compiling and executing both the 'full' and 'noop' kernels (i.e. defining
both EXECFULL and EXECNOOP), the program fails with different error codes, and sometimes it
doesn't fail at all. This happens both on the CPU and the GPU device, although on
the GPU device the program tends to always fail.

The program does not fail when only one kernel is executed (i.e. defining only one
of EXECFULL and EXECNOOP).